[amend] breakoff cache
