+++ common/src/main/kotlin/com/bms/data_attributes/config/impl/AttributeContainerHandler.kt
@@ -1,4 +1,4 @@
-package com.bms.data_attributes.config
+package com.bms.data_attributes.config.impl
 
 import com.bms.data_attributes.config.AttributeConfigManager.Companion.ENTITY_TYPE_INSTANCES
 import com.bms.data_attributes.config.AttributeConfigManager.Tuple
